PycharmProjects/pythonProject5/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
asyncio==3.4.3
cassandra-driver==3.29.1
certifi
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorhash==2.0.0
configparser==5.3.0
dill==0.3.7
Django==3.2.21
dnspython==2.3.0
Flask==2.2.5
Flask-Cors==4.0.1
Flask-MonitoringDashboard==3.3.2
fyers-apiv2==2.0.9
geomet==0.2.1.post1
greenlet==3.0.3
idna==2.10
imbalanced-learn==0.6.1
imblearn==0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.21.6
pandas==1.3.5
pandas-ta==0.3.14b0
Pillow==9.5.0
psutil==5.9.8
psycopg2==2.7.7
pymongo==4.6.1
pyotp==2.9.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.4.4
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tornado==6.1
tqdm==4.66.2
typing_extensions==4.7.1
tzlocal==2.0.0
urllib3==1.26.18
validate-email==1.3
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.2.3
wincertstore==0.2
xgboost==1.6.2
zipp==3.15.0
-e .
